# API-SPECIFICATION.yml
openapi: 3.0.0
paths:
...
  /api/v1/users/profile:
    get:
      operationId: UserController_getProfile
      summary: Get current user profile
      parameters: []
      responses:
        200:
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        401:
          description: Unauthorized
        404:
          description: User not found
      tags:
        - Users
      security:
        -
          JWT-auth: []
    patch:
      operationId: UserController_updateProfile
      summary: Update current user profile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        200:
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        400:
          description: Validation error
        401:
          description: Unauthorized
        404:
          description: User not found
        409:
          description: Phone number already in use
      tags:
        - Users
      security:
        -
          JWT-auth: []
  /api/v1/users/addresses:
    get:
      operationId: UserController_getAddresses
      summary: Get all addresses for current user
      parameters: []
      responses:
        200:
          description: Addresses retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AddressResponseDto"
        401:
          description: Unauthorized
        404:
          description: User not found
      tags:
        - Users
      security:
        -
          JWT-auth: []
    post:
      operationId: UserController_createAddress
      summary: Create a new address
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressDto"
      responses:
        201:
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponseDto"
        400:
          description: Validation error
        401:
          description: Unauthorized
        404:
          description: User not found
      tags:
        - Users
      security:
        -
          JWT-auth: []
  /api/v1/users/addresses/{id}:
    patch:
      operationId: UserController_updateAddress
      summary: Update an address
      parameters:
        -
          name: id
          required: true
          in: path
          description: Address ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddressDto"
      responses:
        200:
          description: Address updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponseDto"
        400:
          description: Validation error
        401:
          description: Unauthorized
        404:
          description: Address not found
      tags:
        - Users
      security:
        -
          JWT-auth: []
    delete:
      operationId: UserController_deleteAddress
      summary: Delete an address
      parameters:
        -
          name: id
          required: true
          in: path
          description: Address ID
          schema:
            type: string
      responses:
        204:
          description: Address deleted
        401:
          description: Unauthorized
        404:
          description: Address not found
      tags:
        - Users
      security:
        -
          JWT-auth: []
  /api/v1/users/addresses/{id}/set-default:
    post:
      operationId: UserController_setDefaultAddress
      summary: Set an address as default
      parameters:
        -
          name: id
          required: true
          in: path
          description: Address ID
          schema:
            type: string
      responses:
        200:
          description: Default address set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponseDto"
        401:
          description: Unauthorized
        404:
          description: Address not found
      tags:
        - Users
      security:
        -
          JWT-auth: []
  /api/v1/restaurants:
    get:
      operationId: RestaurantController_findNearbyRestaurants
      summary: Browse nearby restaurants
      parameters:
        -
          name: latitude
          required: false
          in: query
          description: User latitude for distance calculation
          schema:
            example: 19.076
            type: number
        -
          name: longitude
          required: false
          in: query
          description: User longitude for distance calculation
          schema:
            example: 72.8777
            type: number
        -
          name: radius
          required: false
          in: query
          description: Radius in km
          schema:
            default: 10
            example: 5
            type: number
        -
          name: cuisineType
          required: false
          in: query
          description: Filter by cuisine type
          schema:
            example: Indian
            type: string
        -
          name: isOpen
          required: false
          in: query
          description: Filter only open restaurants
          schema:
            example: true
            type: boolean
        -
          name: search
          required: false
          in: query
          description: Search by restaurant name
          schema:
            example: pizza
            type: string
      responses:
        200:
          description: Restaurants retrieved
        401:
          description: Unauthorized
      tags:
        - Restaurants (Customer)
      security:
        -
          JWT-auth: []
  /api/v1/restaurants/{id}:
    get:
      operationId: RestaurantController_getRestaurantById
      summary: Get restaurant details
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: Restaurant details retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        401:
          description: Unauthorized
        404:
          description: Restaurant not found
      tags:
        - Restaurants (Customer)
      security:
        -
          JWT-auth: []
  /api/v1/restaurants/{id}/menu:
    get:
      operationId: RestaurantController_getRestaurantMenu
      summary: Get restaurant menu
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: string
        -
          name: category
          required: false
          in: query
          schema:
            type: string
      responses:
        200:
          description: Menu retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
        401:
          description: Unauthorized
        404:
          description: Restaurant not found
      tags:
        - Restaurants (Customer)
      security:
        -
          JWT-auth: []
...
info:
  title: SwiftEats API
  description: Real-Time Food Delivery Platform API
  version: 1.0
  contact:
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
            - CUSTOMER
            - RESTAURANT
            - DRIVER
            - ADMIN
        isActive:
          type: boolean
        createdAt:
          format: date-time
          type: string
        lastLoginAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - name
        - phone
        - role
        - isActive
        - createdAt
...
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        phone:
          type: string
          example: +919876543210
    AddressResponseDto:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        label:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        deliveryInstructions:
          type: string
        isDefault:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - userId
        - label
        - street
        - city
        - state
        - postalCode
        - latitude
        - longitude
        - isDefault
        - createdAt
        - updatedAt
    CreateAddressDto:
      type: object
      properties:
        label:
          type: string
          example: Home
        street:
          type: string
          example: 123 Main Street, Apartment 4B
        city:
          type: string
          example: Mumbai
        state:
          type: string
          example: Maharashtra
        postalCode:
          type: string
          example: 400001
        latitude:
          type: number
          example: 19.076
          description: Latitude coordinate (-90 to 90)
        longitude:
          type: number
          example: 72.8777
          description: Longitude coordinate (-180 to 180)
        deliveryInstructions:
          type: string
          example: Apartment 4B, near the park
          default: null
        isDefault:
          type: boolean
          example: false
          default: false
      required:
        - label
        - street
        - city
        - state
        - postalCode
        - latitude
        - longitude
    UpdateAddressDto:
      type: object
      properties:
        label:
          type: string
          example: Home
        street:
          type: string
          example: 123 Main Street, Apartment 4B
        city:
          type: string
          example: Mumbai
        state:
          type: string
          example: Maharashtra
        postalCode:
          type: string
          example: 400001
        latitude:
          type: number
          example: 19.076
          description: Latitude coordinate (-90 to 90)
        longitude:
          type: number
          example: 72.8777
          description: Longitude coordinate (-180 to 180)
        deliveryInstructions:
          type: string
          example: Apartment 4B, near the park
          default: null
        isDefault:
          type: boolean
          example: false
          default: false
    Restaurant:
      type: object
      properties:
    MenuItem:
      type: object
      properties:
...